backend/requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.34.3
sqlalchemy==2.0.47
aiosqlite==0.21.0
greenlet==3.2.3
alembic==1.16.1
pymupdf==1.27.1
ebooklib==0.18
chromadb==1.5.2
openai==2.24.0
tavily-python==0.7.4
cryptography==45.0.3
python-multipart==0.0.20
httpx==0.28.1
pydantic-settings==2.9.1
fastembed
